+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Work around an issue in older lldb-dap binaries where breakpoints would show up as unresolved even when they were hit ([#2283](https://github.com/swiftlang/vscode-swift/pull/2283))
 - Stop assuming that `swift.path` will point at a directory named `bin` ([#2288](https://github.com/swiftlang/vscode-swift/pull/2288))
 - Fixed code coverage information only showing for the first test target when buildnig with swift-build ([#2296](https://github.com/swiftlang/vscode-swift/pull/2296))
+- Fixed test run never completing when stopping a debug session ([#2329](https://github.com/swiftlang/vscode-swift/pull/2329))
 
 ## 2.16.6 - 2026-06-04